publication-scraper-backend/publication_scraper/requirements.txt
anyio==4.4.0
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.7.4
charset-normalizer==3.3.2
comm==0.2.2
crossrefapi==1.6.0
debugpy==1.8.2
decorator==5.1.1
Django==4.2.14
django-filter==24.2
djangorestframework==3.15.2
exceptiongroup==1.2.2
executing==2.0.1
google_search_results==2.4.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.2.0
ipykernel==6.29.5 
ipython==8.18.1
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
Markdown==3.6
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.0.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.0.3
requests==2.32.3
semanticscholar==0.8.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
stack-data==0.6.3
tenacity==8.5.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.16
wcwidth==0.2.13
zipp==3.19.2
colorlog==6.8.2
django-environ==0.11.2
django-filter==24.2
nltk==3.9.1